+++ src/events/messageUpdate.js
@@ -1,5 +1,5 @@
 import { EmbedBuilder } from 'discord.js';
-import { scanImageForScam } from '../utils/antiScam.js';
+import { scanImageForScam, flaggedMessages } from '../utils/antiScam.js';
 import { logToSecurityChannel } from '../utils/helpers.js';
 
 export default {
@@ -26,7 +26,9 @@ export default {
     if (urlsToScan.length > 0) {
       for (const url of urlsToScan) {
          scanImageForScam(url).then(async (isScam) => {
-           if (isScam) {
+           if (isScam && !flaggedMessages.has(newMessage.id)) {
+             flaggedMessages.add(newMessage.id);
+             
              await newMessage.delete().catch(() => null);
              
              // 1. Channel Warning
